# Add ping performance test