--- a/tests/e2e/portal/Kember.js
+++ b/tests/e2e/portal/Kember.js
@@ -74,7 +74,6 @@ async function runTutorial () {
     // check output
     const outFiles2 = [
       "Hear_Rate.csv",
-      "notebooks.zip",
       "Parasympathetic_Cell_Activity.csv",
       "Table_Data.csv"
     ];