--- a/tests/performance/Dockerfile
+++ b/tests/performance/Dockerfile
@@ -1,6 +1,9 @@
 FROM locustio/locust:1.5.3
 
+COPY ./requirements /build/requirements
 RUN pip3 --version && \
   pip3 install \
   faker \
-  python-dotenv
+  python-dotenv \
+  locust-plugins==1.2.0 &&\
+  pip3 freeze --verbose