--- a/tests/performance/Makefile
+++ b/tests/performance/Makefile
@@ -3,6 +3,10 @@
 #
 include ../../scripts/common.Makefile
 
+
+# UTILS
+get_my_ip := $(shell hostname --all-ip-addresses | cut --delimiter=" " --fields=1)
+
 # Check that given variables are set and all have non-empty values,
 # die with an error otherwise.
 #
@@ -16,11 +20,6 @@ __check_defined = \
     $(if $(value $1),, \
       $(error Undefined $1$(if $2, ($2))))
 
-.PHONY: requirements
-requirements: ## compiles pip requirements (.in -> .txt)
-	@$(MAKE_C) requirements reqs
-
-
 .env: .env-devel ## creates .env file from defaults in .env-devel
 	$(if $(wildcard $@), \
 	@echo "WARNING #####  $< is newer than $@ ####"; diff -uN $@ $<; false;,\
@@ -30,28 +29,19 @@ requirements: ## compiles pip requirements (.in -> .txt)
 build:
 	docker buildx build --tag local/locust:latest .
 
-.PHONY: up
+.PHONY: up down run
 up:
 	@$(call check_defined, target, please define target file when calling $@ - e.g. ```make $@ target=MY_LOCUST_FILE.py```)
-	@export TARGET=$(target); \
-	docker-compose --file docker-compose.yml up
+	@export TARGET=$(target); export HOST=$(get_my_ip); \
+	docker-compose --file docker-compose.yml up --scale worker=4
 
 down:
 	docker-compose --file docker-compose.yml down
 
-.PHONY: install
-install-dev: _check_venv_active  ## installs dependencies
-	# installing requirements
-	pip-sync requirements/dev.txt
-
-locusfile=$(if $(target),$(target),locusfile.py)
-
-.PHONY: start
-start: _check_venv_active .env  ## starts locust, a scriptable and scalable performance testing tool
-	# Open http://localhost:8089/
-	locust --locustfile $(locusfile) --host http://127.0.0.1:9081  --users 3 --spawn-rate 1
-
-
-.PHONY: list
-list: _check_venv_active
-	locust --list
+.PHONY: test
+test:
+	@$(call check_defined, target, please define target file when calling $@ - e.g. ```make $@ target=MY_LOCUST_FILE.py```)
+	@export TARGET=$(target); \
+	export HOST=$(get_my_ip); \
+	export LOCUST_OPTIONS="--headless --print-stats --users=100 --spawn-rate=20 --run-time=2m --check-fail-ratio=0.05 --check-avg-response-time=10"; \
+	docker-compose --file docker-compose.yml up --scale worker=4