--- a/tests/performance/docker-compose.yml
+++ b/tests/performance/docker-compose.yml
@@ -6,11 +6,15 @@ services:
     ports:
      - "8089:8089"
     volumes:
-      - ./:/mnt/locust
-    command: -f /mnt/locust/${TARGET} --master -H http://master:8089
+      - ./locust_files:/mnt/locust
+    command: >
+      -f /mnt/locust/${TARGET}
+      --host http://${HOST}:9081
+      ${LOCUST_OPTIONS}
+      --master
 
   worker:
     image: local/locust:latest
     volumes:
-      - ./:/mnt/locust
+      - ./locust_files:/mnt/locust
     command: -f /mnt/locust/${TARGET} --worker --master-host master