--- a/tests/performance/requirements/Makefile
+++ b/tests/performance/requirements/Makefile
@@ -1,6 +1,0 @@
-#
-# Targets to pip-compile requirements
-#
-include ../../../requirements/base.Makefile
-
-# Add here any extra explicit dependency: e.g. _migration.txt: _base.txt