--- a/tests/performance/requirements/_base.txt
+++ b/tests/performance/requirements/_base.txt
@@ -1,6 +1,0 @@
-# NOTE:
-#  This file file is just here as placeholder
-#  to fulfill dependencies of _tools.txt target in requirements/base.Makefile
-#
-#  This is a pure-tests project and all dependencies are added in _test.in
-#