--- a/tests/performance/requirements/_test.in
+++ b/tests/performance/requirements/_test.in
@@ -1,3 +1,0 @@
-locust
-faker
-python-dotenv