--- a/tests/performance/requirements/_tools.in
+++ b/tests/performance/requirements/_tools.in
@@ -1,4 +1,0 @@
--c ../../../requirements/constraints.txt
--c _test.txt
-
--r ../../../requirements/devenv.txt