--- a/tests/performance/requirements/dev.txt
+++ b/tests/performance/requirements/dev.txt
@@ -1,3 +1,0 @@
-# installs base + tests requirements
--r _test.txt
--r _tools.txt