<!-- Common title prefixes/annotations:

  WIP:
  bugfix:
  🏗️ maintenance:

  (⚠️ devops)  = changes in devops config required before deploying
-->

## What do these changes do?
- [x] improves usage of the locust.io perfomance tool
- [x] created an aggressive test for pinging deployments
<!-- Explain REVIEWERS what is this PR about -->

the test simulates 100 users calling 10x the root of the osparc platform and 1x the /v0/health endpoint for 1 minute.
the test fails if the average response time is above 200ms and if the failure rate is above 1%


my current statistics
- local: 0% failure, the average response time: 44ms with a max at 1098ms, doing about 2122 requests/seconds in total.
- master: 0% failure, the average response time: 79ms with a max at 2124ms, doing about 1198 requests/seconds in total.


## Related issue/s

<!-- Enumerate REVIEWERS other issues

e.g.

- #26 : node_ports should have retry policies when upload/download fails  (FIXED)
- ITISFoundation/osparc-issues#304: (Part 2) Prep2Go: creating features to support complex S4L scripts (IMPLEMENTED)

-->


## How to test

<!-- Give REVIEWERS some hits or code snippets on how could this be tested -->


## Checklist

<!-- This is YOUR section

Add here YOUR checklist/notes to guide and monitor the progress of the case!

e.g.

- [ ] Did you change any service's API? Then make sure to bundle document and upgrade version (``make openapi-specs``, ``git commit ...`` and then ``make version-*``)
- [ ] Unit tests for the changes exist
- [ ] Runs in the swarm
- [ ] Documentation reflects the changes
- [ ] New module? Add your github username to [.github/CODEOWNERS](.github/CODEOWNERS)
-->